macosx: new implementation of signal handling using ucontext_t instead of sigcontext+assembly. Now works on MacOS 10.2.1 Betarun version number: 100 -------------------------------------------------------------------------
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@13812 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/cbfa.c
@@ -101,7 +101,7 @@ static void CBFAVisitor(CBEntryFunc func)
 
 void CBFAalloc()
 {
-  DEBUG_CBFA(fprintf(output, "sizeof(CallBackEntry)==%d\n", sizeof(CallBackEntry)));
+  DEBUG_CBFA(fprintf(output, "sizeof(CallBackEntry)==%d\n", (int)sizeof(CallBackEntry)));
 #if 0
 #ifdef linux
   Claim(sizeof(CallBackEntry)==10, "sizeof(CallBackEntry)==10");
